Move window transparency class to graphics
